requirements.txt
apache-airflow[cncf.kubernetes]==2.2.3
